+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2025/05/27
+  * mledit 去掉 scroll_line 属性（感谢俊杰提供补丁)。
+
 2025/05/26
   * 修复 tab\_button\_group\_set\_compact 没有立即生效的问题（感谢俊杰提供补丁)。